refactor: :zap: group tsx and css to folder
