+++ src/configs/url.ts
@@ -1,1 +1,1 @@
-export const BASE_URL = 'http://localhost:8080/'
+export const BASE_URL = 'http://52.63.27.28:8080' //'http://localhost:8080/'
